@@ -20,19 +20,19 @@
     "typescript should be gradually stepped up -- stable as of 2022_09_16"
   ],
   "dependencies": {
-    "@angular/animations": "^13.1.2",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.1.2",
-    "@angular/cli": "^13.1.3",
-    "@angular/common": "^13.1.2",
-    "@angular/compiler": "^13.1.2",
-    "@angular/compiler-cli": "^13.1.2",
-    "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.1.2",
-    "@angular/localize": "^13.1.2",
+    "@angular/cli": "^14.2.10",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
     "@angular/material": "^13.1.2",
-    "@angular/platform-browser": "^13.1.2",
-    "@angular/platform-browser-dynamic": "^13.1.2",
-    "@angular/router": "^13.1.2",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@babel/core": "^7.16.7",
     "@cyber4all/clark-taxonomy": "^4.0.5",
     "aws-sdk": "^2.1058.0",
@@ -54,13 +54,13 @@
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.4",
+    "@angular-devkit/build-angular": "^14.2.10",
     "@angular-eslint/builder": "^13.5.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
     "@angular-eslint/schematics": "^13.0.1",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/language-service": "^13.1.2",
+    "@angular/language-service": "^14.2.12",
     "@types/ckeditor": "^4.9.10",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.12",
