@@ -1,5 +1,5 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { FormGroup } from '@angular/forms';
+import { UntypedFormGroup } from '@angular/forms';
 import { MatchValidator } from 'app/shared/validators/MatchValidator';
 import { AuthValidationService } from 'app/core/auth-validation.service';
 import { AuthService } from 'app/core/auth.service';
@@ -41,7 +41,7 @@ export class ForgotPasswordComponent implements OnInit, OnDestroy{
   showDone: Boolean = false;
   submitButton: String = 'disabled';
 
-  emails: FormGroup = new FormGroup({
+  emails: UntypedFormGroup = new UntypedFormGroup({
     'email': this.authValidationService.getInputFormControl('email'),
     'confirmEmail': this.authValidationService.getInputFormControl('required')
   }, { validators: MatchValidator.mustMatch('email', 'confirmEmail') });;
